create character access function
